+++ src/main/java/joshie/harvest/api/core/Size.java
@@ -1,11 +1,12 @@
 package joshie.harvest.api.core;
 
+import joshie.harvest.api.player.RelationshipType;
 import net.minecraft.util.IStringSerializable;
 
 import java.util.Locale;
 
 public enum Size implements IStringSerializable {
-    SMALL(0), MEDIUM(20000), LARGE(40000);
+    SMALL(0), MEDIUM(RelationshipType.ANIMAL.getMaximumRP() / 2), LARGE(RelationshipType.ANIMAL.getMaximumRP());
 
     private final int relationship;